CRAFTERCMS-427 Fixed bugs causing change template to fail for following use cases: - (content as folder = true) -> (content as folder = false) and vice versa - (file-name) -> (autofile-name) and vice versa Additionally fixed bug with content with autofile-name control not respecting content-as-folder configuration (always creating content as folder)

+++ alfresco-svcs/src/main/java/org/craftercms/cstudio/alfresco/util/impl/PreviewableAspect.java
@@ -146,9 +146,11 @@ public class PreviewableAspect implements NodeServicePolicies.OnAddAspectPolicy,
     @Override
     public void onMoveNode(ChildAssociationRef oldChildAssocRef, final ChildAssociationRef newChildAssocRef) {
         final PersistenceManagerService persistenceManagerService = getServicesManager().getService(PersistenceManagerService.class);
-        String oldParentPath = DmUtils.getNodePath(persistenceManagerService, oldChildAssocRef.getParentRef());
-        String oldName = oldChildAssocRef.getQName().getLocalName();
-        deleteFile(oldParentPath + "/" + oldName);
+        if (oldChildAssocRef.getParentRef() != null && persistenceManagerService.exists(oldChildAssocRef.getParentRef())) {
+            String oldParentPath = DmUtils.getNodePath(persistenceManagerService, oldChildAssocRef.getParentRef());
+            String oldName = oldChildAssocRef.getQName().getLocalName();
+            deleteFile(oldParentPath + "/" + oldName);
+        }
         deployFile(newChildAssocRef.getChildRef());
         final String user = AuthenticationUtil.getFullyAuthenticatedUser();
         final Runnable worker = new Runnable() {
